Fix autogluon.mxnet crash on submodule install
