+++ mxnet/src/autogluon/mxnet/__init__.py
@@ -1,2 +1,1 @@
 from .task import *
-from .task import *
