+++ mxnet/src/autogluon/mxnet/task/__init__.py
@@ -1,6 +1,2 @@
-from .classifier import *
 from .dataset import *
-from .metrics import *
-from .nets import *
-from .classifier import *
 from .utils import *
