+++ mxnet/tests/unittests/test_import.py
@@ -1,0 +1,3 @@
+
+def test_import():
+    import autogluon.mxnet
